+++ source/client/components/App.js
@@ -4,6 +4,7 @@ import Logo from './Logo';
 import TodoList from './todos/TodoList';
 import AddTodoButton from './todos/AddTodoButton';
 
+import 'static/vendor/font-awesome/css/font-awesome.min.css';
 import styles from 'styles/app';
 
 const App = () => (
@@ -28,14 +29,14 @@ const App = () => (
         src="https://ghbtns.com/github-btn.html?user=flipace&repo=lovli.js&type=star&count=true"
         frameBorder="0"
         scrolling="0"
-        width="75px"
+        width="85px"
         height="20px"
       />
       <iframe
         src="https://ghbtns.com/github-btn.html?user=flipace&repo=lovli.js&type=fork&count=true"
         frameBorder="0"
         scrolling="0"
-        width="75px"
+        width="85px"
         height="20px"
       />
     </div>